replace model generator with sdk generator context
