+++ generators/ruby-v2/sdk/src/SdkGeneratorContext.ts
@@ -1,10 +1,13 @@
+import { camelCase, upperFirst } from "lodash-es";
+
 import { GeneratorNotificationService } from "@fern-api/base-generator";
-import { AbstractRubyGeneratorContext } from "@fern-api/ruby-ast";
+import { AbstractRubyGeneratorContext, FileLocation } from "@fern-api/ruby-ast";
 import { RubyProject } from "@fern-api/ruby-base";
 
 import { FernGeneratorExec } from "@fern-fern/generator-exec-sdk";
 import { IntermediateRepresentation } from "@fern-fern/ir-sdk/api";
 
+import { RelativeFilePath } from "../../../../packages/commons/fs-utils/src";
 import { SdkCustomConfigSchema } from "./SdkCustomConfig";
 
 export class SdkGeneratorContext extends AbstractRubyGeneratorContext<SdkCustomConfigSchema> {
@@ -18,5 +21,22 @@ export class SdkGeneratorContext extends AbstractRubyGeneratorContext<SdkCustomC
     ) {
         super(ir, config, customConfig, generatorNotificationService);
         this.project = new RubyProject({ context: this });
+    }
+
+    public getLocationForTypeId(typeId: string): FileLocation {
+        const typeDeclaration = this.ir.types[typeId];
+        if (typeDeclaration == null) {
+            throw new Error(`Type declaration with id ${typeId} not found`);
+        }
+
+        const parts = typeDeclaration.name.fernFilepath.allParts.map((path) => path.pascalCase.safeName);
+        return {
+            namespace: [this.getRootNamespace(), ...parts].join("::"),
+            directory: RelativeFilePath.of(parts.join("/"))
+        };
+    }
+
+    public getRootNamespace(): string {
+        return upperFirst(camelCase(`${this.config.organization}`));
     }
 }
